Add a new POI model as shadow release to extend function of Tags for a postcards
This could be used to point out to a special event or a special place to have e.g. a common set on URLs or even an default description for it.

I've seen that on my collectio there are repeating some events/places/* in a way it makes more sense to add another strcuture aside of tags